// models/measurement.ts
    

    export class Measurement {
        _id: string;
        CustomerId: string;
        ShalwarKameezBottom?: number;
        ShalwarKameezLength?: number;
        ShalwarKameezSLength?: number;
        ShalwarKameezBysep?: number;
        ShalwarKameezNeck?: number;
        ShalwarKameezShoulder?: number;
        ShalwarKameezSleeve?: number;
        ShalwarKameezHip?: number;
        ShalwarKameezWaist?: number;
        ShalwarKameezChest?: number;
        ShalwarKameezMeasurementTakenBy:String;
        __v: number;
        TrouserInside?: number;
        TrouserBottom?: number;
        TrouserKnee?: number;
        TrouserLength?: number;
        TrouserHip?: number;
        TrouserWaist?: number;
        PentInside?: number;
        PentKnee?: number;
        PentLength?: number;
        PentHip?: number;
        PentWaist?: number;
        WaistCoatCB?: number;
        WaistCoatHB?: number;
        WaistCoatNeck?: number;
        WaistCoatShoulder?: number;
        WaistCoatHip?: number;
        WaistCoatWaist?: number;
        WaistCoatChest?: number;
        WaistCoatLength?: number;
        SherwaniBottom?: number;
        SherwaniHB?:number;
        SherwaniCB?: number;
        SherwaniBysep?: number;
        SherwaniNeck?: number;
        SherwaniShoulder?: number;
        SherwaniSleeve?: number;
        SherwaniHip?: number;
        SherwaniWaist?: number;
        SherwaniChest?: number;
        SherwaniLength?: number;
        CoatCB?: number;
        CoatHB?: number;
        CoatBysep?: number;
        CoatNeck?: number;
        CoatShoulder?: number;
        CoatSleeve?: number;
        CoatHip?: number;
        CoatWaist?: number;
        CoatChest?: number;
        CoatMeasurementTakenBy:String;    
        SherwaniMeasurementTakenBy:String; 
        WaistCoatMeasurementTakenBy:String;    
        PentMeasurementTakenBy:String;
        TrouserMeasurementTakenBy:String;
        ShawlLength:string;
        ShawlWidth:string;
    }